+++ HidingBar/HidingBar.lua
@@ -37,8 +37,14 @@ local ignoreFrameList = {
 
 
 local function void() end
-local function enter() hidingBar:enter() end
-local function leave() hidingBar:leave() end
+local function enter()
+	hidingBar.isMouse = true
+	hidingBar:enter()
+end
+local function leave()
+	hidingBar.isMouse = false
+	hidingBar:leave()
+end
 
 
 if MSQ then
@@ -172,8 +178,13 @@ if MSQ then
 			local normal = btn:GetNormalTexture()
 			if normal then
 				icon = btn:CreateTexture(nil, "BACKGROUND")
-				icon:SetTexture(normal:GetTexture())
-				icon:SetTexCoord(normal:GetTexCoord())
+				local atlas = normal:GetAtlas()
+				if atlas then
+					icon:SetAtlas(atlas)
+				else
+					icon:SetTexture(normal:GetTexture())
+					icon:SetTexCoord(normal:GetTexCoord())
+				end
 				icon:SetVertexColor(normal:GetVertexColor())
 				icon:SetSize(normal:GetSize())
 				for i = 1, normal:GetNumPoints() do
@@ -477,19 +488,16 @@ do
 				button.icon:SetDesaturated(data.iconDesaturated)
 			end
 		end
-		if data.OnClick then
-			button:SetScript("OnClick", data.OnClick)
-		end
 		button:HookScript("OnEnter", enter)
 		button:HookScript("OnLeave", leave)
 		button.IsShown = IsShown
 		tinsert(self.createdButtons, button)
 		tinsert(self.mixedButtons, button)
+
 		if update then
 			self:sort()
-			self:applyLayout()
-			self:leave()
-			self.cb:Fire("BUTTON_ADDED", name, button, update)
+			self:setButtonSize()
+			self.cb:Fire("BUTTON_ADDED", name, button, true)
 		end
 
 		if self.MSQ_Button then
@@ -1255,7 +1263,6 @@ end)
 
 
 function hidingBar:enter(force)
-	self.isMouse = true
 	if not self.isDrag and self.shown and (self.config.showHandler ~= 3 or force) then
 		UIFrameFadeRemoveFrame(self.drag)
 		self.drag:SetAlpha(1)
@@ -1282,7 +1289,6 @@ end
 
 
 function hidingBar:leave(timer)
-	self.isMouse = false
 	if not self.isDrag and self:IsShown() and self.config.showHandler ~= 3 then
 		self.timer = timer or self.config.hideDelay
 		self:SetScript("OnUpdate", self.hideBar)
@@ -1299,7 +1305,7 @@ function hidingBar:refreshShown()
 		self.drag:Hide()
 		if self.config.showHandler == 3 then
 			self:enter(true)
-		elseif self:IsShown() then
+		elseif self:IsShown() and not self.isMouse then
 			self:leave()
 		end
 	elseif self.config.showHandler == 3 then
@@ -1307,7 +1313,9 @@ function hidingBar:refreshShown()
 		self.drag:SetShown(not self.config.lock)
 	else
 		if self:IsShown() then
-			self:leave()
+			if not self.isMouse then
+				self:leave()
+			end
 		else
 			self:updateDragBarPosition()
 			if self.config.fade then