+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # HidingBar - Classic
 
-* Added localization for Germany, thx SlayerEGT
+* Fixed hiding of the bar when adding a button when the bar is displayed.
+* Added Korean localization.